+++ src/components/Header/Header.module.css
@@ -71,7 +71,7 @@
   text-align: left;
 }
 
-.cainButton {
+.chainButton {
   position: relative;
   align-items: right;
   margin-top: 9px;
