--- src/components/Header/index.module.css
@@ -1,42 +1,0 @@
-.header {
-  display: flex;
-  flex-direction: row;
-  font-size: 16px;
-  color: #e2fb5f;
-  font-family: "Bai Jamjuree", sans-serif;
-  justify-content: flex-start;
-}
-
-.hero {
-  display: flex;
-  flex-direction: column;
-  justify-self: start;
-}
-
-.description {
-  width: 590px;
-  text-align: left;
-  color: #43b25b;
-  font-family: "Bai Jamjuree", sans-serif;
-  margin-top: 23px;
-}
-
-.link {
-  color: #e2fb5f;
-  align-self: left;
-}
-
-.linkhead {
-  color: #e2fb5f;
-  text-decoration: none;
-  font-size: 46px;
-  font-family: "Bai Jamjuree", sans-serif;
-  font-weight: 700;
-  justify-self: start;
-}
-
-.header .walletWrapper {
-  justify-self: end;
-  margin-top: 19px;
-  margin-left: auto;
-}
